Restore Bill of Rights #6 and #7 to their original language.
These two lines were said to be too restrictive, but that was intended to be the point.

If the purpose of government is to protect person and property, then those two lines do exactly that.

Every consitution that has ever been written has failed to limit government. If historical precendent holds, this one will fail as well. But we make the language as restrictive as we possibly can then maybe we stand a change.

Watering down the language to make it easier for the government to create laws will just expidite the erosion of the constitution.