fair representation
The U.S. Constitution provides for 1 representative for every 30,000 people. Unfortunately, Congress froze the size of the House of Representatives in the early 20th century. As a result, the people's vote for their representative has been steadily diluted over the past 100 years. Today there is 1 representative in the USA for every 700,000 people.

Let's be clear in understanding that 1 representative for 10,000 people is fair representation, whereas 1 representative for a million people is tyranny. To protect against this "tyranny from dilution" we must insure a minimum number of representatives for every X number of citizens.