3/4s of the Assembly
Since fifteen is 3/4s of twenty, I have changed the wording from fifteen to three-quarters. (Effectively the same thing, except it also works regardless of the actual number of assembly representatives.) This way if the number of Assembly members grows to meet the needs of fair representation, the 3/4s requirement will stand strong regardless.