right to bear arms
The U.S. Constitution protects the right of the people to keep and bear arms. Nevertheless, there are many infringements upon and violations of this right.

Often we see people arrested simply for having a gun in their car, or while moving their household from one place to another. We also see licensing requirements used to restrict gun ownership. Often when people apply for a license, it is not granted, or the terms are too restrictive to make gun ownership practically available.

The changes in this pull request are designed to ensure the government absolutely cannot infringe upon the right of the people to bear arms, based on problems that have historically occurred in the USA.