Fixed "majority" to "majority of all"
The original text read that there are "20" representatives, and that "11" or "15" must vote to do certain things.

I changed "11" and "15" to "a majority" and "three-quarters". Unfortunately, I did not specify that it was "a majority of ALL Assembly members" and "three-quarters of ALL Assembly members."

This created a danger that a "majority" could occur with less than 11, if some of the Assembly members failed to participate in the vote.

Therefore I have fixed this problem with this PR, in order to retain the original intent.