eleven/fifteen
I changed "vote of eleven" to be "vote of a majority of all Assembly Representatives."

I also changed "vote of fifteen" to be "vote of three-quarters of all Assembly Representatives."

NOTE: this PR does *not* change the operation of the Constitution at all. **The behavior is identical.** The purpose of this PR is to make it easier later to make changes in §IV.2 with a future pull request. This PR stands on its own, and since it doesn't change the functionality of the Constitution in any way, it should be fine to merge it. This way any future debate can occur whenever a PR is submitted for §IV.2, and it won't require changes all over the Constitution.