Section II. 7 further changes to consider
No section in the constitution prevents the government from transferring property from one person to another. As of now, section II.7 only addresses situations where the government confiscates property for the public good. The changes I made fix that, but it should be noted that this will also make any cancellation of debt illegal since it entails the transfer of property.     

>II.7. No law shall allow for confiscation of private property, either fully or partially, for any purpose other than reparation or the collection of evidence in pursuant to a Warrant; nor shall it regulate the use of property, so long as no negative rights of others are infringed; anything found in proximity under one’s land or growing naturally on it shall belong to oneself.

Any thoughts?