Rewording
With the previous wording you could interpret it so that each assembly rep.'s pay can't exceed 5% of the Annual Budget, not the total of all A.R.s. I also think the total percent of the budget should be higher, but that doesn't have to be discussed at this time.