Removal of ARs
In Liberland an AR needs only 5% of the population to cast their ballots in favour of them to gain a seat in the Assembly. It makes no sense to have any clause for the public to be able to impeach this Representative. What I see happening is that people who didn't vote for this AR just vote to impeach him thus giving majorities are larger representation in the government.